requirements.txt
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
billiard==3.6.4.0
blis==0.7.9
catalogue==2.0.8
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
confection==0.0.4
cryptography==39.0.0
cymem==2.0.7
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
fastapi==0.89.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
kombu==5.2.4
langcodes==3.3.0
MarkupSafe==2.1.2
murmurhash==1.0.9
numpy==1.24.1
packaging==23.0
passlib==1.7.4
pathy==0.10.1
preshed==3.0.8
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.7.1
redis==4.4.2
requests==2.28.2
rsa==4.9
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.0
SQLAlchemy-Utils==0.39.0
srsly==2.4.5
starlette==0.22.0
thinc==8.1.7
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
vine==5.0.0
wasabi==1.1.1
wcwidth==0.2.6

asgiref~=3.6.0
bcrypt~=4.0.1